Issue/8

+++ www/model/cart.php
@@ -1,6 +1,7 @@
 <?php 
 require_once MODEL_PATH . 'functions.php';
 require_once MODEL_PATH . 'db.php';
+require_once MODEL_PATH . 'order.php';
 
 function get_user_carts($db, $user_id){
   $sql = "
@@ -21,9 +22,9 @@ function get_user_carts($db, $user_id){
     ON
       carts.item_id = items.item_id
     WHERE
-      carts.user_id = {$user_id}
+      carts.user_id = ?
   ";
-  return fetch_all_query($db, $sql);
+  return fetch_all_query($db, $sql,array($user_id));
 }
 
 function get_user_cart($db, $user_id, $item_id){
@@ -45,12 +46,12 @@ function get_user_cart($db, $user_id, $item_id){
     ON
       carts.item_id = items.item_id
     WHERE
-      carts.user_id = {$user_id}
+      carts.user_id = ?
     AND
-      items.item_id = {$item_id}
+      items.item_id = ?
   ";
 
-  return fetch_query($db, $sql);
+  return fetch_query($db, $sql, array($user_id, $item_id));
 
 }
 
@@ -70,10 +71,10 @@ function insert_cart($db, $user_id, $item_id, $amount = 1){
         user_id,
         amount
       )
-    VALUES({$item_id}, {$user_id}, {$amount})
+    VALUES(?, ?, ?)
   ";
 
-  return execute_query($db, $sql);
+  return execute_query($db, $sql, array($item_id, $user_id, $amount));
 }
 
 function update_cart_amount($db, $cart_id, $amount){
@@ -81,12 +82,12 @@ function update_cart_amount($db, $cart_id, $amount){
     UPDATE
       carts
     SET
-      amount = {$amount}
+      amount = ?
     WHERE
-      cart_id = {$cart_id}
+      cart_id = ?
     LIMIT 1
   ";
-  return execute_query($db, $sql);
+  return execute_query($db, $sql, array($amount, $cart_id));
 }
 
 function delete_cart($db, $cart_id){
@@ -94,39 +95,46 @@ function delete_cart($db, $cart_id){
     DELETE FROM
       carts
     WHERE
-      cart_id = {$cart_id}
+      cart_id = ?
     LIMIT 1
   ";
 
-  return execute_query($db, $sql);
+  return execute_query($db, $sql, array($cart_id));
 }
 
 function purchase_carts($db, $carts){
   if(validate_cart_purchase($carts) === false){
     return false;
   }
-  foreach($carts as $cart){
-    if(update_item_stock(
-        $db, 
-        $cart['item_id'], 
-        $cart['stock'] - $cart['amount']
-      ) === false){
-      set_error($cart['name'] . 'の購入に失敗しました。');
+  $db->beginTransaction();
+  if(insert_orders($db, $carts[0]['user_id']) !== false){
+    $order_id = $db->lastInsertId('order_id');
+    foreach($carts as $cart){
+      if(update_item_stock($db, $cart['item_id'], $cart['stock'] - $cart['amount'])
+        && insert_orders_detail($db, $order_id, $cart['item_id'], $cart['amount'], $cart['price']) === false)
+      {
+        set_error($cart['name'] . 'の購入に失敗しました。');
+        $db->rollback();
+        return false;
+      }
     }
+    $db->commit();
+    delete_user_carts($db, $carts[0]['user_id']);
+  }else{
+    return false;
   }
-  
-  delete_user_carts($db, $carts[0]['user_id']);
 }
+
 
 function delete_user_carts($db, $user_id){
   $sql = "
     DELETE FROM
       carts
     WHERE
-      user_id = {$user_id}
+      user_id = ?
   ";
 
-  execute_query($db, $sql);
+  execute_query($db, $sql, array($user_id));
 }
 
 
@@ -156,4 +164,3 @@ function validate_cart_purchase($carts){
   }
   return true;
 }
-
